Move keyboard_keys to utils folder
