+++ src/epomakercontroller/keyboard_gui.py
@@ -2,7 +2,7 @@ from pathlib import Path
 import tkinter as tk
 from tkinter.colorchooser import askcolor as askcolour   # thats right
 
-from .keyboard_keys import KeyboardKey, KeyboardKeys
+from .utils.keyboard_keys import KeyboardKey, KeyboardKeys
 from .commands.EpomakerKeyRGBCommand import KeyMap, KeyboardRGBFrame
 from typing import Callable, Literal
 from .configs.configs import Config
